backend/requirements.txt
supabase
flask
flask_cors
flask_sqlalchemy
python-dotenv
requests
rapidfuzz
gtts
SpeechRecognition
pydub
numpy
opencv-python
Pillow